Refractor file operations into file_manager
